+++ src/extension.ts
@@ -1,7 +1,6 @@
 import * as vscode from "vscode";
 import S4TKWorkspace from "#workspace/s4tk-workspace";
 import registerContributions from "./contributions/_index";
-import { indexWorkspace } from "#workspace/indexing";
 
 let _extensionContext: vscode.ExtensionContext;
 export const getExtensionContext = () => _extensionContext;
@@ -10,5 +9,4 @@ export function activate(context: vscode.ExtensionContext) {
 	_extensionContext = context;
 	registerContributions(context);
 	S4TKWorkspace.activate();
-	indexWorkspace();
 }